[Aashish]ADDED:Test for two plus one inch compare with three inch

+++ QuantityMeasurementSln/Quantity.cs
@@ -14,7 +14,7 @@ namespace QuantityMeasurementSln
             conversionUnit = new ConversionUnit();
         }
 
-        public Quantity(int value, string unit)
+        public Quantity(double value, string unit)
         {
             this.value = value;
             this.unit = unit;
